refactor(workflow-runtime): 시뮬레이션 개선 후보 승인 책임을 분리
